Refactored operators. (#532)
